requirements.txt
anthropic
python-telegram-bot
httpx
duckduckgo-search
wikipedia
google-auth-oauthlib
google-api-python-client
pypdf
